refactor(billing): move PromoCodeService to services directory and remove unused utils
